workaround for complex jax.random.normal precision bug

+++ Test/Optimizer/test_sr_onthefly_logic.py
@@ -62,7 +62,10 @@ def random_split_like_tree(rng_key, target=None, treedef=None):
 def tree_random_normal_like(rng_key, target):
     keys_tree = random_split_like_tree(rng_key, target)
     return jax.tree_multimap(
-        lambda l, k: jax.random.normal(k, l.shape, l.dtype), target, keys_tree
+        # TODO remove astype once its fixed in jax
+        lambda l, k: jax.random.normal(k, l.shape, l.dtype).astype(l.dtype),
+        target,
+        keys_tree,
     )
